Modularized Tokenizer
