+++ sqltidy/tokenizer/__init__.py
@@ -1,0 +1,22 @@
+# tokenizer/__init__.py
+"""
+Tokenizer package init: exposes main API for tokenization and grouping.
+"""
+from .core import tokenize, tokenize_with_types, TokenType, Token, SemanticLevel, TokenGroup, GroupType
+from .grouping import group_parentheses, group_by_statements, group_by_clauses, group_tokens
+from .patterns import apply_patterns
+
+__all__ = [
+    "tokenize",
+    "tokenize_with_types",
+    "TokenType",
+    "Token",
+    "SemanticLevel",
+    "TokenGroup",
+    "GroupType",
+    "group_parentheses",
+    "group_by_statements",
+    "group_by_clauses",
+    "group_tokens",
+    "apply_patterns",
+]
