+++ sqltidy/tokenizer/grouping.py
@@ -1,0 +1,24 @@
+# tokenizer/grouping.py
+"""
+Token grouping logic: parentheses, statements, clauses, and general grouping helpers.
+"""
+from typing import List, Union
+from .core import Token, TokenType
+from ..dialects import SQLDialect
+
+class GroupType:
+    STATEMENT = "statement"
+    CLAUSE = "clause"
+    PARENTHESIS = "parenthesis"
+    FUNCTION = "function"
+    # ... add other group types as needed ...
+
+class TokenGroup:
+    def __init__(self, group_type, tokens, name=None, metadata=None):
+        self.group_type = group_type
+        self.tokens = tokens
+        self.name = name
+        self.metadata = metadata or {}
+
+# ...existing code for group_parentheses, group_by_statements, group_by_clauses, group_tokens...
+# For brevity, you can copy the grouping functions from tokenizer.py and adapt imports as needed.
