+++ sqltidy/tokenizer/patterns.py
@@ -1,0 +1,10 @@
+# tokenizer/patterns.py
+"""
+Advanced pattern recognition for semantic SQL constructs (CASE, window functions, etc).
+"""
+# This is a stub. Move or refactor pattern_tokenizer logic here.
+
+def apply_patterns(tokens, dialect):
+    # Import or implement advanced pattern matching logic here
+    # For now, just return tokens unchanged
+    return tokens
